refactor: address code review feedback
- use better Go package names
- break out Go packages more granularly
- make some functions private
